requirements.txt
flask>=3.0.0
flask-jwt-extended>=4.6.0
pymysql>=1.1.0
qrcode[pil]>=7.4.2
pillow>=10.0.0
python-dotenv>=1.0.0
# Optional but recommended for production-grade connection pooling:
# DBUtils>=3.0.0
